+++ apps/wallet-mobile/src/yoroi-wallets/hooks/index.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import {Certificate} from '@emurgo/cross-csl-core'
 import AsyncStorage, {AsyncStorageStatic} from '@react-native-async-storage/async-storage'
 import {useNavigation} from '@react-navigation/native'
 import {parseBoolean, useMutationWithInvalidations, useStorage} from '@yoroi/common'
@@ -1089,4 +1090,13 @@ export const useNativeAssetInvalidation = ({networkId, policy, name}: NativeAsse
     ...mutation,
     invalidate: mutation.mutate,
   }
+}
+
+export const useCreateGovernanceTx = (wallet: YoroiWallet) => {
+  const mutationFn = (certificates: Certificate[]) => {
+    return wallet.createUnsignedGovernanceTx(certificates)
+  }
+
+  const mutation = useMutation({mutationFn, useErrorBoundary: true, retry: false})
+  return {...mutation, createUnsignedGovernanceTx: mutation.mutateAsync}
 }
